fix data display with new database

+++ database/seeds/ProductTableSeeder.php
@@ -4,6 +4,7 @@ use Illuminate\Database\Seeder;
 use Faker\Factory as Faker;
 use Carbon\Carbon;
 use App\Model\Product;
+use App\Model\ProductDetail;
 use App\Model\ProductElastic;
 use Elasticsearch\ClientBuilder;
 use App\Repositories\Product\ProductRepositoryInterface;
@@ -25,24 +26,25 @@ class ProductTableSeeder extends Seeder
             'index' => $productElastic->getIndexName()
         ];
         $client = ClientBuilder::create()->build();
-        $client->indices()->delete($param);
+        // check index exists before delete
+        if ($client->indices()->exists($param)) {
+            $client->indices()->delete($param);
+        }
 
         // add data to mongo db product table
     	$productRepository = app(ProductRepositoryInterface::class);
         $faker = Faker::create('fr');
         for ($i = 0; $i < 100; $i++) { 
         	$title = $faker->text(20);
-        	$alias = str_slug($title);
-        	$shortDescription = $faker->text(30);
-        	$description = $faker->text(200);
+            $url = str_slug($title);
+        	$description = $faker->text(30);
         	$image = '/image/front/Bibliotheque_Web_1.jpg';
-        	$price = 1;
-        	$views = 2;
+        	$view = 2;
         	$userId = 1;
-        	$catId = 1;
+        	$categoryId = 1;
         	$like = 0;
-        	$status = 0;
-        	$share = 0;
+            $share = 0;
+        	$pink = 0;
         	$isPublic = 0;
         	$isDelete = 0;
         	$createdAt = Carbon::now()->format('Y-m-d H:i:s');
@@ -50,55 +52,70 @@ class ProductTableSeeder extends Seeder
 
             $data = [
                 'title'             => $title,
-                'alias'             => $alias,
-                'short_description'  => $shortDescription,
+                'url'               => $url,
                 'description'       => $description,
                 'image'             => $image,
-                'price'             => $price,
-                'views'             => $views,
-                'user_id'           	=> $userId,
-                'cat_id'           	=> $catId,
-                'like'           	=> $like,
-                'status'           	=> $status,
-                'share'           	=> $share,
-                'is_public'         => $isPublic,
+                'view'              => $view,
+                'like'              => $like,
+                'category_id'       => $categoryId,
                 'is_delete'         => $isDelete,
                 'created_at'        => $createdAt,
                 'updated_at'        => $updatedAt,
             ];
 
-            Product::create($data);
+            $productCreate = Product::create($data);
+            $dataProductDetail = [
+                'product_id' => $productCreate->id,
+                'user_id' => $userId,
+                'share' => $share,
+                'pink' => $pink,
+                'is_public' => $isPublic,
+                'is_delete' => $isDelete,
+            ];
+
+            ProductDetail::create($dataProductDetail);
         }
 
         // insert data to Elastic
         $products = $productRepository->all();
        	if (count($products) > 0) {
        		foreach ($products as $product) {
-       			$dataElastic = [
-                    'body' => [
-                        'title'             => $product->title,
-		                'alias'             => $product->alias,
-		                'short_description' => $product->short_description,
-		                'description'       => $product->description,
-		                'image'             => $product->image,
-		                'price'             => $product->price,
-		                'views'             => $product->views,
-		                'user_id'           => $product->user_id,
-		                'cat_id'           	=> $product->cat_id,
-		                'like'           	=> $product->like,
-		                'status'           	=> $product->status,
-		                'share'           	=> $product->share,
-		                'is_public'         => $product->is_public,
-		                'is_delete'         => $product->is_delete,
-		                'created_at'        => $product->created_at,
-		                'updated_at'        => $product->updated_at,
-                    ],
-                    'index' => $productElastic->getIndexName(),
-                    'type' => $productElastic->getTypeName(),
-                    'id' => $product->id,
-                ];
-                $client = ClientBuilder::create()->build();
-                $client->index($dataElastic);
+                $productDetail = ProductDetail::where('product_id', $product->id)->first();
+                if ($productDetail) {
+                    $productDetailElatic = [
+                        'id'            => $productDetail->id,
+                        'product_id'    => $product->id,
+                        'user_id'       => $productDetail->user_id,
+                        'share'         => $productDetail->share,
+                        'pink'          => $productDetail->pink,
+                        'is_public'     => $productDetail->is_public,
+                        'is_delete'     => $productDetail->is_delete,
+                        'created_at'    => $productDetail->created_at,
+                        'updated_at'    => $productDetail->updated_at,
+                    ];
+
+                    $dataElastic = [
+                        'body' => [
+                            'title'             => $product->title,
+                            'url'               => $product->url,
+                            'description'       => $product->description,
+                            'image'             => $product->image,
+                            'view'              => $product->view,
+                            'category_id'       => $product->categoryId_id,
+                            'like'              => $product->like,
+                            'is_delete'         => $product->is_delete,
+                            'created_at'        => $product->created_at,
+                            'updated_at'        => $product->updated_at,
+                            'product_detail'    => $productDetailElatic
+                        ],
+                        'index' => $productElastic->getIndexName(),
+                        'type' => $productElastic->getTypeName(),
+                        'id' => $product->id,
+                    ];
+
+                    $client = ClientBuilder::create()->build();
+                    $client->index($dataElastic);
+                }
        		}
        	}
     }
